add load\save annota file funcion

+++ FrameModel.m
@@ -14,14 +14,35 @@ classdef FrameModel < handle
         end
         
         function bbObj = getObj(obj, bbId)
+            bbObj = [];
+            if bbId <= 0 || obj.m_bbMap.Count == 0
+                return;
+            end   
            bbObj = obj.m_bbMap(num2str(bbId));
         end
         
         function removeObj(obj, bbId)
-            bbObj = obj.m_bbMap(num2str(bbId));
-            bbObj.delete();
-%            bbObj.deleteRect();
+            if obj.m_bbMap.Count <= 0
+                return;
+            end
+            if ~isKey(obj.m_bbMap,num2str(bbId))
+                return;
+            end
+             bbObj = obj.m_bbMap(num2str(bbId));
+             bbObj.delete();
             remove(obj.m_bbMap, num2str(bbId));
+        end
+        
+        function removeAllbb(obj)
+            if obj.m_bbMap.Count <= 0
+                return;
+            end
+            keySet = keys(obj.m_bbMap);
+            len = length(keySet);
+            for i = 1:len
+               strId = keySet{i};
+               obj.removeObj(strId);
+            end
         end
         
     end
